move batched diag
